requirements.txt
torch>=2.0
torchvision
matplotlib
captum
opentelemetry-sdk

